+++ rpc-rs/src/common.rs
@@ -1,3 +1,4 @@
+use crate::RpcError;
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
@@ -12,18 +13,14 @@ pub struct Message<'m> {
     pub arg: Cow<'m, [u8]>,
 }
 
-/// context data
-pub mod context {
+/// Context - message passing metadata used by wasmhost Actors and Capability Providers
+#[derive(Default, Debug, Clone)]
+pub struct Context {
+    /// Messages received by Context Provider will have actor set to the actor's public key
+    pub actor: Option<String>,
 
-    /// Context - message passing metadata used by wasmhost Actors and Capability Providers
-    #[derive(Default, Debug, Clone)]
-    pub struct Context {
-        /// Messages received by Context Provider will have actor set to the actor's public key
-        pub actor: Option<String>,
-
-        /// Span name/context for tracing. This is a placeholder for now
-        pub span: Option<String>,
-    }
+    /// Span name/context for tracing. This is a placeholder for now
+    pub span: Option<String>,
 }
 
 /// Client config defines the intended recipient of a message and parameters that transport may use to adapt sending it
@@ -54,7 +51,7 @@ impl SendOpts {
 pub trait Transport: Send {
     async fn send(
         &self,
-        ctx: &context::Context,
+        ctx: &Context,
         req: Message<'_>,
         opts: Option<SendOpts>,
     ) -> std::result::Result<Vec<u8>, RpcError>;
@@ -86,99 +83,7 @@ cfg_if::cfg_if! {
     }
 }
 
-/// An error that can occur in the processing of an RPC. This is not request-specific errors but
-/// rather cross-cutting errors that can always occur.
-#[derive(thiserror::Error, Debug, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum RpcError {
-    /// The request exceeded its deadline.
-    #[error("the request exceeded its deadline: {0}")]
-    DeadlineExceeded(String),
-
-    /// A capability provider was called before its configure_dispatch was called.
-    #[error("the capability provider has not been initialized: {0}")]
-    NotInitialized(String),
-
-    #[error("method not handled {0}")]
-    MethodNotHandled(String),
-
-    /// Error that can be returned if server has not implemented
-    /// an optional interface method
-    #[error("method not implemented")]
-    NotImplemented,
-
-    #[error("Host send error {0}")]
-    HostError(String),
-
-    #[error("deserialization: {0}")]
-    Deser(String),
-
-    #[error("serialization: {0}")]
-    Ser(String),
-
-    #[error("rpc: {0}")]
-    Rpc(String),
-
-    #[error("nats: {0}")]
-    Nats(String),
-
-    #[error("invalid parameter: {0}")]
-    InvalidParameter(String),
-
-    /// Error occurred in actor's rpc handler
-    #[error("actor: {0}")]
-    ActorHandler(String),
-
-    /// Error occurred during provider initialization or put-link
-    #[error("provider initialization or put-link: {0}")]
-    ProviderInit(String),
-
-    /// Timeout occurred
-    #[error("timeout: {0}")]
-    Timeout(String),
-
-    //#[error("IO error")]
-    //IO([from] std::io::Error)
-    /// Anything else
-    #[error("{0}")]
-    Other(String),
-}
-
-impl From<String> for RpcError {
-    fn from(s: String) -> RpcError {
-        RpcError::Other(s)
-    }
-}
-
-impl From<&str> for RpcError {
-    fn from(s: &str) -> RpcError {
-        RpcError::Other(s.to_string())
-    }
-}
-
-impl From<std::io::Error> for RpcError {
-    fn from(e: std::io::Error) -> RpcError {
-        RpcError::Other(format!("io: {}", e))
-    }
-}
-
-impl From<minicbor::encode::Error<std::io::Error>> for RpcError {
-    fn from(e: minicbor::encode::Error<std::io::Error>) -> RpcError {
-        RpcError::Ser(format!("cbor-encode: {}", e))
-    }
-}
-
-impl From<minicbor::decode::Error> for RpcError {
-    fn from(e: minicbor::decode::Error) -> RpcError {
-        RpcError::Deser(format!("cbor-decode: {}", e))
-    }
-}
-
 #[async_trait]
 pub trait MessageDispatch {
-    async fn dispatch(
-        &self,
-        ctx: &context::Context,
-        message: Message<'_>,
-    ) -> Result<Message<'_>, RpcError>;
+    async fn dispatch(&self, ctx: &Context, message: Message<'_>) -> Result<Message<'_>, RpcError>;
 }
