+++ rpc-rs/src/lib.rs
@@ -5,14 +5,12 @@
 //!
 
 mod timestamp;
+// re-export Timestamp
 pub use timestamp::Timestamp;
 
 mod actor_wasm;
-mod common;
-pub use common::{
-    context::Context, deserialize, serialize, Message, MessageDispatch, RpcError, SendOpts,
-    Transport,
-};
+pub mod common;
+use common::{deserialize, serialize, Context, Message, MessageDispatch, SendOpts, Transport};
 pub mod channel_log;
 pub mod provider;
 pub(crate) mod provider_main;
@@ -234,8 +232,9 @@ pub mod actor {
 
     pub mod prelude {
         pub use crate::{
+            common::{Context, Message, MessageDispatch, Transport},
             core::{Actor, ActorReceiver},
-            Context, Message, MessageDispatch, RpcError, RpcResult, Transport,
+            RpcError, RpcResult,
         };
 
         // re-export async_trait
@@ -273,5 +272,93 @@ pub mod actor {
                 pub fn console_log(_s: &str) {}
             }
         }
+    }
+}
+
+/// An error that can occur in the processing of an RPC. This is not request-specific errors but
+/// rather cross-cutting errors that can always occur.
+#[derive(thiserror::Error, Debug, serde::Serialize, serde::Deserialize)]
+#[non_exhaustive]
+pub enum RpcError {
+    /// The request exceeded its deadline.
+    #[error("the request exceeded its deadline: {0}")]
+    DeadlineExceeded(String),
+
+    /// A capability provider was called before its configure_dispatch was called.
+    #[error("the capability provider has not been initialized: {0}")]
+    NotInitialized(String),
+
+    #[error("method not handled {0}")]
+    MethodNotHandled(String),
+
+    /// Error that can be returned if server has not implemented
+    /// an optional interface method
+    #[error("method not implemented")]
+    NotImplemented,
+
+    #[error("Host send error {0}")]
+    HostError(String),
+
+    #[error("deserialization: {0}")]
+    Deser(String),
+
+    #[error("serialization: {0}")]
+    Ser(String),
+
+    #[error("rpc: {0}")]
+    Rpc(String),
+
+    #[error("nats: {0}")]
+    Nats(String),
+
+    #[error("invalid parameter: {0}")]
+    InvalidParameter(String),
+
+    /// Error occurred in actor's rpc handler
+    #[error("actor: {0}")]
+    ActorHandler(String),
+
+    /// Error occurred during provider initialization or put-link
+    #[error("provider initialization or put-link: {0}")]
+    ProviderInit(String),
+
+    /// Timeout occurred
+    #[error("timeout: {0}")]
+    Timeout(String),
+
+    //#[error("IO error")]
+    //IO([from] std::io::Error)
+    /// Anything else
+    #[error("{0}")]
+    Other(String),
+}
+
+impl From<String> for RpcError {
+    fn from(s: String) -> RpcError {
+        RpcError::Other(s)
+    }
+}
+
+impl From<&str> for RpcError {
+    fn from(s: &str) -> RpcError {
+        RpcError::Other(s.to_string())
+    }
+}
+
+impl From<std::io::Error> for RpcError {
+    fn from(e: std::io::Error) -> RpcError {
+        RpcError::Other(format!("io: {}", e))
+    }
+}
+
+impl From<minicbor::encode::Error<std::io::Error>> for RpcError {
+    fn from(e: minicbor::encode::Error<std::io::Error>) -> RpcError {
+        RpcError::Ser(format!("cbor-encode: {}", e))
+    }
+}
+
+impl From<minicbor::decode::Error> for RpcError {
+    fn from(e: minicbor::decode::Error) -> RpcError {
+        RpcError::Deser(format!("cbor-decode: {}", e))
     }
 }
