+++ rpc-rs/src/provider.rs
@@ -43,15 +43,15 @@ trait ProviderImpl: ProviderDispatch + Send + Sync + Clone + 'static {}
 
 pub mod prelude {
     pub use crate::{
+        common::{Context, Message, MessageDispatch, SendOpts},
         core::LinkDefinition,
         provider::{HostBridge, NatsClient, ProviderDispatch, ProviderHandler},
         provider_main::{
             get_host_bridge, load_host_data, provider_main, provider_run, provider_start,
         },
-        Context, Message, MessageDispatch, RpcError, RpcResult, SendOpts,
+        RpcError, RpcResult,
     };
 
-    //pub use crate::Timestamp;
     pub use async_trait::async_trait;
     pub use wasmbus_macros::Provider;
