+++ rpc-rs/src/wasmbus_model.rs
@@ -1,7 +1,7 @@
 // This file is generated automatically using wasmcloud/weld-codegen and smithy model definitions
 //
 
-#![allow(dead_code, unused_imports, clippy::ptr_arg, clippy::needless_lifetimes)]
+#[allow(unused_imports)]
 use serde::{Deserialize, Serialize};
 
 pub const SMITHY_VERSION: &str = "1.0";
